==> scripts/0_chargement_packages.R <==
library(sf)
library(dplyr)
library(tidyverse)
library(tmaptools)
library(ggplot2)
library(ggthemes)
library(patchwork)
library(readr)
library(viridis)
library(classInt)
library(plotly)
library(fuzzyjoin)
library(ggfx)
library(ggnewscale)
library(fmsb)
library(reshape2)
library(treemap)
library(flextable)
library(scales)
library(stringi)
library(officer)
library(magick)
